feat(tmq): update table list wher alter tag & split big files in tmq
