+++ source/client/src/clientRawBlockWrite.c
@@ -16,6 +16,7 @@
 #include <string.h>
 #include "cJSON.h"
 #include "clientInt.h"
+#include "clientRawBlockWrite.h"
 #include "nodes.h"
 #include "osMemPool.h"
 #include "osMemory.h"
@@ -31,74 +32,6 @@
 #include "tglobal.h"
 #include "tmsgtype.h"
 
-#define RAW_LOG_END                                                           \
-  if (code != 0) {                                                            \
-    uError("%s failed at line:%d since:%s", __func__, lino, tstrerror(code)); \
-  } else {                                                                    \
-    uDebug("%s return success", __func__);                                    \
-  }
-
-#define RAW_LOG_START uDebug("%s start", __func__);
-
-#define RAW_NULL_CHECK(c) \
-  do {                    \
-    if (c == NULL) {      \
-      lino = __LINE__;    \
-      code = terrno;      \
-      goto end;           \
-    }                     \
-  } while (0)
-
-#define RAW_FALSE_CHECK(c)           \
-  do {                               \
-    if (!(c)) {                      \
-      code = TSDB_CODE_INVALID_PARA; \
-      lino = __LINE__;               \
-      goto end;                      \
-    }                                \
-  } while (0)
-
-#define RAW_RETURN_CHECK(c) \
-  do {                      \
-    code = c;               \
-    if (code != 0) {        \
-      lino = __LINE__;      \
-      goto end;             \
-    }                       \
-  } while (0)
-
-#define LOG_ID_TAG   "connId:0x%" PRIx64 ", QID:0x%" PRIx64 ",func:%s"
-#define LOG_ID_VALUE *(int64_t*)taos, pRequest->requestId, __func__
-
-#define TMQ_META_VERSION "1.0"
-
-static cJSON* tmqAddObjectToArray(cJSON* array) {
-  cJSON* item = cJSON_CreateObject();
-  if (cJSON_AddItemToArray(array, item)) {
-    return item;
-  }
-  cJSON_Delete(item);
-  return NULL;
-}
-
-static cJSON* tmqAddStringToArray(cJSON* array, const char* str) {
-  cJSON* item = cJSON_CreateString(str);
-  if (cJSON_AddItemToArray(array, item)) {
-    return item;
-  }
-  cJSON_Delete(item);
-  return NULL;
-}
- 
-#define ADD_TO_JSON_STRING(JSON,NAME,VALUE) \
-  RAW_NULL_CHECK(cJSON_AddStringToObject(JSON, NAME, VALUE));
-
-#define ADD_TO_JSON_BOOL(JSON,NAME,VALUE) \
-  RAW_NULL_CHECK(cJSON_AddBoolToObject(JSON, NAME, VALUE));
-
-#define ADD_TO_JSON_NUMBER(JSON,NAME,VALUE) \
-  RAW_NULL_CHECK(cJSON_AddNumberToObject(JSON, NAME, VALUE));
-
 static int32_t  tmqWriteBatchMetaDataImpl(TAOS* taos, void* meta, uint32_t metaLen);
 static tb_uid_t processSuid(tb_uid_t suid, char* db) {
   if (db == NULL) {
@@ -106,903 +39,6 @@ static tb_uid_t processSuid(tb_uid_t suid, char* db) {
   }
   return suid + MurmurHash3_32(db, strlen(db));
 }
-
-static int32_t getLength(int8_t type, int32_t bytes, int32_t typeMod) {
-  int32_t length = 0;
-  if (type == TSDB_DATA_TYPE_BINARY || type == TSDB_DATA_TYPE_VARBINARY || type == TSDB_DATA_TYPE_GEOMETRY) {
-    length = bytes - VARSTR_HEADER_SIZE;
-  } else if (type == TSDB_DATA_TYPE_NCHAR) {
-    length = (bytes - VARSTR_HEADER_SIZE) / TSDB_NCHAR_SIZE;
-  } else if (IS_STR_DATA_BLOB(type)) {
-    length = bytes - BLOBSTR_HEADER_SIZE;
-  } else if (type == TSDB_DATA_TYPE_DECIMAL || type == TSDB_DATA_TYPE_DECIMAL64) {
-    length = typeMod;
-  }
-  return length;
-}
-
-static int32_t buildCreateTableJson(SSchemaWrapper* schemaRow, SSchemaWrapper* schemaTag, SExtSchema* pExtSchemas, char* name, int64_t id, int8_t t,
-                                 bool isVirtual, SColRefWrapper* colRef, SColCmprWrapper* pColCmprRow, cJSON** pJson) {
-  if (schemaRow == NULL || name == NULL || pColCmprRow == NULL || pJson == NULL) {
-    uError("invalid parameter, schemaRow:%p, name:%p, pColCmprRow:%p, pJson:%p", schemaRow, name, pColCmprRow, pJson);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t code = TSDB_CODE_SUCCESS;
-  int32_t lino = 0;
-  int8_t  buildDefaultCompress = 0;
-  if (pColCmprRow->nCols <= 0) {
-    buildDefaultCompress = 1;
-  }
-  RAW_LOG_START
-  char*  string = NULL;
-  cJSON* json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-
-  ADD_TO_JSON_STRING(json, "type", "create");
-  ADD_TO_JSON_STRING(json, "tableType", (t == TSDB_SUPER_TABLE ? "super" : "normal"));
-  if (isVirtual){
-    ADD_TO_JSON_BOOL(json, "isVirtual", isVirtual);
-  }
-  ADD_TO_JSON_STRING(json, "tableName", name);
-
-  cJSON* columns = cJSON_AddArrayToObject(json, "columns");
-  RAW_NULL_CHECK(columns);
-
-  for (int i = 0; i < schemaRow->nCols; i++) {
-    cJSON* column = tmqAddObjectToArray(columns);
-    RAW_NULL_CHECK(column);
-    SSchema* s = schemaRow->pSchema + i;
-    ADD_TO_JSON_STRING(column, "name", s->name);
-    ADD_TO_JSON_NUMBER(column, "type", s->type);
-    int32_t typeMod = 0;
-    if (pExtSchemas != NULL) {
-      typeMod = pExtSchemas[i].typeMod;
-    }
-    int32_t length = getLength(s->type, s->bytes, typeMod);
-    if (length > 0) {
-      ADD_TO_JSON_NUMBER(column, "length", length);
-    }
-
-    if (isVirtual && colRef != NULL && i < colRef->nCols){
-      SColRef* pColRef = colRef->pColRef + i;
-      if (pColRef->hasRef) {
-        cJSON* ref = cJSON_AddObjectToObject(column, "ref");
-        RAW_NULL_CHECK(ref);
-        ADD_TO_JSON_STRING(ref, "refDbName", pColRef->refDbName);
-        ADD_TO_JSON_STRING(ref, "refTableName", pColRef->refTableName);
-        ADD_TO_JSON_STRING(ref, "refColName", pColRef->refColName);
-      }
-    }
-    ADD_TO_JSON_BOOL(column, "isPrimarykey", (s->flags & COL_IS_KEY));
-    if (pColCmprRow == NULL) {
-      continue;
-    }
-
-    uint32_t alg = 0;
-    if (buildDefaultCompress) {
-      alg = createDefaultColCmprByType(s->type);
-    } else {
-      SColCmpr* pColCmpr = pColCmprRow->pColCmpr + i;
-      alg = pColCmpr->alg;
-    }
-    const char* encode = columnEncodeStr(COMPRESS_L1_TYPE_U32(alg));
-    RAW_NULL_CHECK(encode);
-    const char* compress = columnCompressStr(COMPRESS_L2_TYPE_U32(alg));
-    RAW_NULL_CHECK(compress);
-    const char* level = columnLevelStr(COMPRESS_L2_TYPE_LEVEL_U32(alg));
-    RAW_NULL_CHECK(level);
-
-    ADD_TO_JSON_STRING(column, "encode", encode);
-    ADD_TO_JSON_STRING(column, "compress", compress);
-    ADD_TO_JSON_STRING(column, "level", level);
-  }
-
-  cJSON* tags = cJSON_AddArrayToObject(json, "tags");
-  RAW_NULL_CHECK(tags);
-
-  for (int i = 0; schemaTag && i < schemaTag->nCols; i++) {
-    cJSON* tag = tmqAddObjectToArray(tags);
-    RAW_NULL_CHECK(tag);
-    SSchema* s = schemaTag->pSchema + i;
-    ADD_TO_JSON_STRING(tag, "name", s->name);
-    ADD_TO_JSON_NUMBER(tag, "type", s->type);
-    int32_t length = getLength(s->type, s->bytes, 0);
-    if (length > 0) {
-      ADD_TO_JSON_NUMBER(tag, "length", length);
-    }
-  }
-
-end:
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t setCompressOption(cJSON* json, uint32_t para) {
-  if (json == NULL) {
-    return TSDB_CODE_INVALID_PARA;
-  }
-  uint8_t encode = COMPRESS_L1_TYPE_U32(para);
-  int32_t code = 0;
-  int32_t lino = 0;
-  RAW_LOG_START
-  if (encode != 0) {
-    const char* encodeStr = columnEncodeStr(encode);
-    RAW_NULL_CHECK(encodeStr);
-    ADD_TO_JSON_STRING(json, "encode", encodeStr);
-    goto end;
-  }
-  uint8_t compress = COMPRESS_L2_TYPE_U32(para);
-  if (compress != 0) {
-    const char* compressStr = columnCompressStr(compress);
-    RAW_NULL_CHECK(compressStr);
-    ADD_TO_JSON_STRING(json, "compress", compressStr);
-    goto end;
-  }
-  uint8_t level = COMPRESS_L2_TYPE_LEVEL_U32(para);
-  if (level != 0) {
-    const char* levelStr = columnLevelStr(level);
-    RAW_NULL_CHECK(levelStr);
-    ADD_TO_JSON_STRING(json, "level", levelStr);
-    goto end;
-  }
-
-end:
-  RAW_LOG_END
-  return code;
-}
-static int32_t buildAlterSTableJson(void* alterData, int32_t alterDataLen, cJSON** pJson) {
-  if (alterData == NULL || pJson == NULL) {
-    uError("invalid parameter in %s alterData:%p", __func__, alterData);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SMAlterStbReq req = {0};
-  cJSON*        json = NULL;
-  char*         string = NULL;
-  int32_t       code = 0;
-  int32_t       lino = 0;
-  RAW_LOG_START
-  RAW_RETURN_CHECK(tDeserializeSMAlterStbReq(alterData, alterDataLen, &req));
-  json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-  ADD_TO_JSON_STRING(json, "type", "alter");
-  SName name = {0};
-  RAW_RETURN_CHECK(tNameFromString(&name, req.name, T_NAME_ACCT | T_NAME_DB | T_NAME_TABLE));
-  ADD_TO_JSON_STRING(json, "tableType", "super");
-  ADD_TO_JSON_STRING(json, "tableName", name.tname);
-  ADD_TO_JSON_NUMBER(json, "alterType", req.alterType);
-
-  switch (req.alterType) {
-    case TSDB_ALTER_TABLE_ADD_TAG:
-    case TSDB_ALTER_TABLE_ADD_COLUMN: {
-      if (taosArrayGetSize(req.pFields) != 1) {
-        uError("invalid field num %" PRIzu " for alter type %d", taosArrayGetSize(req.pFields), req.alterType);
-        cJSON_Delete(json);
-        json = NULL;
-        code = TSDB_CODE_INVALID_PARA;
-        goto end;
-      }
-      TAOS_FIELD* field = taosArrayGet(req.pFields, 0);
-      RAW_NULL_CHECK(field);
-      ADD_TO_JSON_STRING(json, "colName", field->name);
-      ADD_TO_JSON_NUMBER(json, "colType", field->type);
-      int32_t typeMode = 0;
-      if (taosArrayGetSize(req.pTypeMods) > 0) {
-        typeMode = *(STypeMod*)taosArrayGet(req.pTypeMods, 0);
-      }
-      int32_t length = getLength(field->type, field->bytes, typeMode);
-      if (length > 0) {
-        ADD_TO_JSON_NUMBER(json, "colLength", length);
-      }
-
-      break;
-    }
-    case TSDB_ALTER_TABLE_ADD_COLUMN_WITH_COMPRESS_OPTION: {
-      SFieldWithOptions* field = taosArrayGet(req.pFields, 0);
-      RAW_NULL_CHECK(field);
-      ADD_TO_JSON_STRING(json, "colName", field->name);
-      ADD_TO_JSON_NUMBER(json, "colType", field->type);
-      int32_t typeMode = 0;
-      if (taosArrayGetSize(req.pTypeMods) > 0) {
-        typeMode = *(STypeMod*)taosArrayGet(req.pTypeMods, 0);
-      }
-      int32_t length = getLength(field->type, field->bytes, typeMode);
-      if (length > 0) {
-        ADD_TO_JSON_NUMBER(json, "colLength", length);
-      }
-
-      RAW_RETURN_CHECK(setCompressOption(json, field->compress));
-      break;
-    }
-    case TSDB_ALTER_TABLE_DROP_TAG:
-    case TSDB_ALTER_TABLE_DROP_COLUMN: {
-      TAOS_FIELD* field = taosArrayGet(req.pFields, 0);
-      RAW_NULL_CHECK(field);
-      ADD_TO_JSON_STRING(json, "colName", field->name);
-      break;
-    }
-    case TSDB_ALTER_TABLE_UPDATE_TAG_BYTES:
-    case TSDB_ALTER_TABLE_UPDATE_COLUMN_BYTES: {
-      if (taosArrayGetSize(req.pFields) != 1) {
-        uError("invalid field num %" PRIzu " for alter type %d", taosArrayGetSize(req.pFields), req.alterType);
-        cJSON_Delete(json);
-        json = NULL;
-        code = TSDB_CODE_INVALID_PARA;
-        goto end;
-      }
-      TAOS_FIELD* field = taosArrayGet(req.pFields, 0);
-      RAW_NULL_CHECK(field);
-      ADD_TO_JSON_STRING(json, "colName", field->name);
-      ADD_TO_JSON_NUMBER(json, "colType", field->type);
-      int32_t length = getLength(field->type, field->bytes, 0);
-      if (length > 0) {
-        ADD_TO_JSON_NUMBER(json, "colLength", length);
-      }
-
-      break;
-    }
-    case TSDB_ALTER_TABLE_UPDATE_TAG_NAME:
-    case TSDB_ALTER_TABLE_UPDATE_COLUMN_NAME: {
-      TAOS_FIELD* oldField = taosArrayGet(req.pFields, 0);
-      RAW_NULL_CHECK(oldField);
-      TAOS_FIELD* newField = taosArrayGet(req.pFields, 1);
-      RAW_NULL_CHECK(newField);
-      ADD_TO_JSON_STRING(json, "colName", oldField->name);
-      ADD_TO_JSON_STRING(json, "colNewName", newField->name);
-      break;
-    }
-    case TSDB_ALTER_TABLE_UPDATE_COLUMN_COMPRESS: {
-      TAOS_FIELD* field = taosArrayGet(req.pFields, 0);
-      RAW_NULL_CHECK(field);
-      ADD_TO_JSON_STRING(json, "colName", field->name);
-      RAW_RETURN_CHECK(setCompressOption(json, field->bytes));
-      break;
-    }
-    default:
-      break;
-  }
-
-end:
-  tFreeSMAltertbReq(&req);
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processCreateStb(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (metaRsp == NULL || pJson == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t        code = TSDB_CODE_SUCCESS;
-  int32_t        lino = 0;
-  SVCreateStbReq req = {0};
-  SDecoder       coder = {0};
-
-  RAW_LOG_START
-  // decode and process req
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-  tDecoderInit(&coder, data, len);
-
-  RAW_RETURN_CHECK(tDecodeSVCreateStbReq(&coder, &req));
-  RAW_RETURN_CHECK(buildCreateTableJson(&req.schemaRow, &req.schemaTag, req.pExtSchemas, req.name, req.suid,
-                                        TSDB_SUPER_TABLE, req.virtualStb, NULL, &req.colCmpr, pJson));
-
-end:
-  tDecoderClear(&coder);
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processAlterStb(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (metaRsp == NULL || pJson == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SVCreateStbReq req = {0};
-  SDecoder       coder = {0};
-  int32_t        code = TSDB_CODE_SUCCESS;
-  int32_t        lino = 0;
-  RAW_LOG_START
-
-  // decode and process req
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-  tDecoderInit(&coder, data, len);
-
-  RAW_RETURN_CHECK(tDecodeSVCreateStbReq(&coder, &req));
-  RAW_RETURN_CHECK(buildAlterSTableJson(req.alterOriData, req.alterOriDataLen, pJson));
-
-end:
-  tDecoderClear(&coder);
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t buildChildElement(cJSON* json, SVCreateTbReq* pCreateReq) {
-  if (json == NULL || pCreateReq == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  STag*   pTag = (STag*)pCreateReq->ctb.pTag;
-  char*   sname = pCreateReq->ctb.stbName;
-  char*   name = pCreateReq->name;
-  SArray* tagName = pCreateReq->ctb.tagName;
-  int64_t id = pCreateReq->uid;
-  uint8_t tagNum = pCreateReq->ctb.tagNum;
-  int32_t code = 0;
-  int32_t lino = 0;
-  SArray* pTagVals = NULL;
-  char*   pJson = NULL;
-  char*   buf = NULL;
-  RAW_LOG_START
-
-  ADD_TO_JSON_STRING(json, "tableName", name);
-
-  if (pCreateReq->type == TSDB_VIRTUAL_CHILD_TABLE) {
-    cJSON* refs = cJSON_AddArrayToObject(json, "refs");
-    RAW_NULL_CHECK(refs);
-
-    for (int i = 0; i < pCreateReq->colRef.nCols; i++) {
-      SColRef* pColRef = pCreateReq->colRef.pColRef + i;
-  
-      if (!pColRef->hasRef) {
-        continue;
-      }
-      cJSON* ref = tmqAddObjectToArray(refs);
-      RAW_NULL_CHECK(ref);
-      ADD_TO_JSON_STRING(ref, "colName", pColRef->colName);
-      ADD_TO_JSON_STRING(ref, "refDbName", pColRef->refDbName);
-      ADD_TO_JSON_STRING(ref, "refTableName", pColRef->refTableName);
-      ADD_TO_JSON_STRING(ref, "refColName", pColRef->refColName);
-    }
-  }
-
-  ADD_TO_JSON_STRING(json, "using", sname);
-  ADD_TO_JSON_NUMBER(json, "tagNum", tagNum);
-
-  cJSON* tags = cJSON_AddArrayToObject(json, "tags");
-  RAW_NULL_CHECK(tags);
-  RAW_RETURN_CHECK(tTagToValArray(pTag, &pTagVals));
-  if (tTagIsJson(pTag)) {
-    STag* p = (STag*)pTag;
-    if (p->nTag == 0) {
-      uWarn("p->nTag == 0");
-      goto end;
-    }
-    parseTagDatatoJson(pTag, &pJson, NULL);
-    RAW_NULL_CHECK(pJson);
-    cJSON* tag = tmqAddObjectToArray(tags);
-    RAW_NULL_CHECK(tag);
-    STagVal* pTagVal = taosArrayGet(pTagVals, 0);
-    RAW_NULL_CHECK(pTagVal);
-    char* ptname = taosArrayGet(tagName, 0);
-    RAW_NULL_CHECK(ptname);
-    ADD_TO_JSON_STRING(tag, "name", ptname);
-    ADD_TO_JSON_NUMBER(tag, "type", TSDB_DATA_TYPE_JSON);
-    ADD_TO_JSON_STRING(tag, "value", pJson);
-    goto end;
-  }
-
-  for (int i = 0; i < taosArrayGetSize(pTagVals); i++) {
-    STagVal* pTagVal = (STagVal*)taosArrayGet(pTagVals, i);
-    RAW_NULL_CHECK(pTagVal);
-    cJSON* tag = tmqAddObjectToArray(tags);
-    RAW_NULL_CHECK(tag);
-    char* ptname = taosArrayGet(tagName, i);
-    RAW_NULL_CHECK(ptname);
-    ADD_TO_JSON_STRING(tag, "name", ptname);
-    ADD_TO_JSON_NUMBER(tag, "type", pTagVal->type);
-
-    if (IS_VAR_DATA_TYPE(pTagVal->type)) {
-      if (IS_STR_DATA_BLOB(pTagVal->type)) {
-        goto end;
-      }
-      int64_t bufSize = 0;
-      if (pTagVal->type == TSDB_DATA_TYPE_VARBINARY) {
-        bufSize = pTagVal->nData * 2 + 2 + 3;
-      } else {
-        bufSize = pTagVal->nData + 3;
-      }
-      buf = taosMemoryCalloc(bufSize, 1);
-      RAW_NULL_CHECK(buf);
-      code = dataConverToStr(buf, bufSize, pTagVal->type, pTagVal->pData, pTagVal->nData, NULL);
-      if (code != TSDB_CODE_SUCCESS) {
-        uError("convert tag value to string failed");
-        goto end;
-      }
-
-      ADD_TO_JSON_STRING(tag, "value", buf)
-      taosMemoryFreeClear(buf);
-    } else {
-      double val = 0;
-      GET_TYPED_DATA(val, double, pTagVal->type, &pTagVal->i64, 0);  // currently tag type can't be decimal, so pass 0 as typeMod
-      ADD_TO_JSON_NUMBER(tag, "value", val)
-    }
-  }
-
-end:
-  taosMemoryFree(pJson);
-  taosArrayDestroy(pTagVals);
-  taosMemoryFree(buf);
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t buildCreateCTableJson(SVCreateTbReq* pCreateReq, int32_t nReqs, cJSON** pJson) {
-  if (pJson == NULL || pCreateReq == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t code = 0;
-  int32_t lino = 0;
-  RAW_LOG_START
-  char*  string = NULL;
-  cJSON* json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-  ADD_TO_JSON_STRING(json, "type", "create");
-  ADD_TO_JSON_STRING(json, "tableType", "child");
-
-  RAW_RETURN_CHECK(buildChildElement(json, pCreateReq));
-  cJSON* createList = cJSON_AddArrayToObject(json, "createList");
-  RAW_NULL_CHECK(createList);
-
-  for (int i = 0; nReqs > 1 && i < nReqs; i++) {
-    cJSON* create = tmqAddObjectToArray(createList);
-    RAW_NULL_CHECK(create);
-    RAW_RETURN_CHECK(buildChildElement(create, pCreateReq + i));
-  }
-
-end:
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processCreateTable(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (pJson == NULL || metaRsp == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t            code = TSDB_CODE_SUCCESS;
-  int32_t            lino = 0;
-  SDecoder           decoder = {0};
-  SVCreateTbBatchReq req = {0};
-  SVCreateTbReq*     pCreateReq;
-  RAW_LOG_START
-  // decode
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-  tDecoderInit(&decoder, data, len);
-  RAW_RETURN_CHECK(tDecodeSVCreateTbBatchReq(&decoder, &req));
-  // loop to create table
-  if (req.nReqs > 0) {
-    pCreateReq = req.pReqs;
-    if (pCreateReq->type == TSDB_CHILD_TABLE || pCreateReq->type == TSDB_VIRTUAL_CHILD_TABLE) {
-      RAW_RETURN_CHECK(buildCreateCTableJson(req.pReqs, req.nReqs, pJson));
-    } else if (pCreateReq->type == TSDB_NORMAL_TABLE || pCreateReq->type == TSDB_VIRTUAL_NORMAL_TABLE) {
-      RAW_RETURN_CHECK(buildCreateTableJson(&pCreateReq->ntb.schemaRow, NULL, pCreateReq->pExtSchemas, pCreateReq->name,
-                                            pCreateReq->uid, pCreateReq->type, pCreateReq->type == TSDB_NORMAL_TABLE ? false : true, 
-                                            &pCreateReq->colRef, &pCreateReq->colCmpr, pJson));
-    }
-  }
-
-end:
-  tDeleteSVCreateTbBatchReq(&req);
-  tDecoderClear(&decoder);
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processAutoCreateTable(SMqDataRsp* rsp, char** string) {
-  int32_t lino = 0;
-  int32_t code = TSDB_CODE_SUCCESS;
-  RAW_LOG_START
-  RAW_FALSE_CHECK(rsp != NULL && string != NULL);
-  SDecoder*      decoder = NULL;
-  SVCreateTbReq* pCreateReq = NULL;
-  RAW_FALSE_CHECK(rsp->createTableNum > 0);
-
-  decoder = taosMemoryCalloc(rsp->createTableNum, sizeof(SDecoder));
-  RAW_NULL_CHECK(decoder);
-  pCreateReq = taosMemoryCalloc(rsp->createTableNum, sizeof(SVCreateTbReq));
-  RAW_NULL_CHECK(pCreateReq);
-
-  // loop to create table
-  for (int32_t iReq = 0; iReq < rsp->createTableNum; iReq++) {
-    // decode
-    void** data = taosArrayGet(rsp->createTableReq, iReq);
-    RAW_NULL_CHECK(data);
-    int32_t* len = taosArrayGet(rsp->createTableLen, iReq);
-    RAW_NULL_CHECK(len);
-    tDecoderInit(&decoder[iReq], *data, *len);
-    RAW_RETURN_CHECK(tDecodeSVCreateTbReq(&decoder[iReq], pCreateReq + iReq));
-
-    if (pCreateReq[iReq].type != TSDB_CHILD_TABLE && pCreateReq[iReq].type != TSDB_NORMAL_TABLE) {
-      uError("%s failed. pCreateReq[iReq].type:%d invalid", __func__, pCreateReq[iReq].type);
-      code = TSDB_CODE_INVALID_PARA;
-      goto end;
-    }
-  }
-  cJSON* pJson = NULL;
-  if (pCreateReq->type == TSDB_NORMAL_TABLE) {
-    RAW_RETURN_CHECK(buildCreateTableJson(&pCreateReq->ntb.schemaRow, NULL, pCreateReq->pExtSchemas, pCreateReq->name,
-                                          pCreateReq->uid, TSDB_NORMAL_TABLE, false, NULL, &pCreateReq->colCmpr, &pJson));
-  } else if (pCreateReq->type == TSDB_CHILD_TABLE) {
-    RAW_RETURN_CHECK(buildCreateCTableJson(pCreateReq, rsp->createTableNum, &pJson));
-  }
-
-  *string = cJSON_PrintUnformatted(pJson);
-  cJSON_Delete(pJson);
-
-  uDebug("auto created table return, sql json:%s", *string);
-
-end:
-  RAW_LOG_END
-  for (int i = 0; decoder && pCreateReq && i < rsp->createTableNum; i++) {
-    tDecoderClear(&decoder[i]);
-    taosMemoryFreeClear(pCreateReq[i].comment);
-    if (pCreateReq[i].type == TSDB_CHILD_TABLE) {
-      taosArrayDestroy(pCreateReq[i].ctb.tagName);
-    }
-  }
-  taosMemoryFree(decoder);
-  taosMemoryFree(pCreateReq);
-  return code;
-}
-
-static int32_t processAlterTable(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (pJson == NULL || metaRsp == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SDecoder     decoder = {0};
-  SVAlterTbReq vAlterTbReq = {0};
-  char*        string = NULL;
-  cJSON*       json = NULL;
-  int32_t      code = 0;
-  int32_t      lino = 0;
-  char*        buf = NULL;
-  char*        buf1 = NULL;
-  char*        buf2 = NULL;
-  SNode*       pWhere = NULL;
-
-  RAW_LOG_START
-
-  // decode
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-  tDecoderInit(&decoder, data, len);
-  RAW_RETURN_CHECK(tDecodeSVAlterTbReq(&decoder, &vAlterTbReq));
-
-  json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-  ADD_TO_JSON_STRING(json, "type", "alter");
-
-  char* tableType = NULL;
-  if (vAlterTbReq.action == TSDB_ALTER_TABLE_UPDATE_MULTI_TABLE_TAG_VAL){
-    tableType = "child";
-  } else if (vAlterTbReq.action == TSDB_ALTER_TABLE_UPDATE_CHILD_TABLE_TAG_VAL){
-    tableType = "super";
-  } else if (vAlterTbReq.action == TSDB_ALTER_TABLE_ALTER_COLUMN_REF ||
-             vAlterTbReq.action == TSDB_ALTER_TABLE_REMOVE_COLUMN_REF) {
-    tableType = "";
-  } else {
-    tableType = "normal";
-  }
-
-  ADD_TO_JSON_STRING(json, "tableType", tableType);
-  ADD_TO_JSON_STRING(json, "tableName", vAlterTbReq.action == TSDB_ALTER_TABLE_UPDATE_MULTI_TABLE_TAG_VAL ? "" : vAlterTbReq.tbName);
-  ADD_TO_JSON_NUMBER(json, "alterType", vAlterTbReq.action);
-
-  uDebug("alter table action:%d", vAlterTbReq.action);
-  switch (vAlterTbReq.action) {
-    case TSDB_ALTER_TABLE_ADD_COLUMN: 
-    case TSDB_ALTER_TABLE_ADD_COLUMN_WITH_COLUMN_REF: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      ADD_TO_JSON_NUMBER(json, "colType", vAlterTbReq.type);
-
-      int32_t length = getLength(vAlterTbReq.type, vAlterTbReq.bytes, vAlterTbReq.typeMod);
-      if (length > 0) {
-        ADD_TO_JSON_NUMBER(json, "colLength", length);
-      }
-
-      if (vAlterTbReq.action == TSDB_ALTER_TABLE_ADD_COLUMN_WITH_COLUMN_REF) {
-        ADD_TO_JSON_STRING(json, "refDbName", vAlterTbReq.refDbName);
-        ADD_TO_JSON_STRING(json, "refTbName", vAlterTbReq.refTbName);
-        ADD_TO_JSON_STRING(json, "refColName", vAlterTbReq.refColName);
-      }
-      break;
-    }
-    case TSDB_ALTER_TABLE_ADD_COLUMN_WITH_COMPRESS_OPTION: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      ADD_TO_JSON_NUMBER(json, "colType", vAlterTbReq.type);
-
-      int32_t length = getLength(vAlterTbReq.type, vAlterTbReq.bytes, vAlterTbReq.typeMod);
-      if (length > 0) {
-        ADD_TO_JSON_NUMBER(json, "colLength", length);
-      }
-      RAW_RETURN_CHECK(setCompressOption(json, vAlterTbReq.compress));
-      break;
-    }
-    case TSDB_ALTER_TABLE_DROP_COLUMN: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      break;
-    }
-    case TSDB_ALTER_TABLE_UPDATE_COLUMN_BYTES: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      ADD_TO_JSON_NUMBER(json, "colType", vAlterTbReq.colModType);
-      int32_t length = getLength(vAlterTbReq.colModType, vAlterTbReq.colModBytes, vAlterTbReq.typeMod);
-      if (length > 0) {
-        ADD_TO_JSON_NUMBER(json, "colLength", length);
-      }
-      break;
-    }
-    case TSDB_ALTER_TABLE_UPDATE_COLUMN_NAME: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      ADD_TO_JSON_STRING(json, "colNewName", vAlterTbReq.colNewName);
-      break;
-    }
-    case TSDB_ALTER_TABLE_UPDATE_MULTI_TABLE_TAG_VAL: {
-      int32_t nTables = taosArrayGetSize(vAlterTbReq.tables);
-      if (nTables <= 0) {
-        code = TSDB_CODE_INVALID_PARA;
-        uError("processAlterTable parse multi tables error");
-        goto end;
-      }
-
-      cJSON* tables = cJSON_AddArrayToObject(json, "tables");
-      RAW_NULL_CHECK(tables);
-
-      for (int32_t i = 0; i < nTables; i++) {
-        SUpdateTableTagVal* pTable = taosArrayGet(vAlterTbReq.tables, i);
-        cJSON* tableObj = tmqAddObjectToArray(tables);
-        RAW_NULL_CHECK(tableObj);
-
-        ADD_TO_JSON_STRING(tableObj, "tableName", pTable->tbName);
-
-        int32_t nTags = taosArrayGetSize(pTable->tags);
-        cJSON* tags = cJSON_AddArrayToObject(tableObj, "tags");
-        RAW_NULL_CHECK(tags);
-
-        for (int32_t j = 0; j < nTags; j++) {
-          cJSON* member = tmqAddObjectToArray(tags);
-          RAW_NULL_CHECK(member);
-
-          SUpdatedTagVal* pTagVal = taosArrayGet(pTable->tags, j);
-          ADD_TO_JSON_STRING(member, "colName", pTagVal->tagName);
-
-          if (pTagVal->regexp != NULL) {
-            ADD_TO_JSON_STRING(member, "regexp", pTagVal->regexp);
-            ADD_TO_JSON_STRING(member, "replacement", pTagVal->replacement);
-          } else {
-            bool isNull = pTagVal->isNull;
-            if (!isNull) {
-              int64_t bufSize = 0;
-              if (pTagVal->tagType == TSDB_DATA_TYPE_VARBINARY) {
-                bufSize = pTagVal->nTagVal * 2 + 2 + 3;
-              } else {
-                bufSize = pTagVal->nTagVal + 3;
-              }
-              buf1 = taosMemoryCalloc(bufSize, 1);
-              RAW_NULL_CHECK(buf1);
-              code = dataConverToStr(buf1, bufSize, pTagVal->tagType, pTagVal->pTagVal, pTagVal->nTagVal, NULL);
-              if (code != TSDB_CODE_SUCCESS) {
-                uError("convert tag value to string failed");
-                goto end;
-              }
-              ADD_TO_JSON_STRING(member, "colValue", buf1);
-              taosMemoryFreeClear(buf1);
-            }
-            ADD_TO_JSON_BOOL(member, "colValueNull", isNull);
-          }
-        }
-      }
-      break;
-    }
-
-    case TSDB_ALTER_TABLE_UPDATE_CHILD_TABLE_TAG_VAL: {
-      int32_t nTags = taosArrayGetSize(vAlterTbReq.pMultiTag);
-      if (nTags <= 0) {
-        code = TSDB_CODE_INVALID_PARA;
-        uError("processAlterTable parse multi tags error");
-        goto end;
-      }
-
-      cJSON* tags = cJSON_AddArrayToObject(json, "tags");
-      RAW_NULL_CHECK(tags);
-
-      for (int32_t i = 0; i < nTags; i++) {
-        cJSON* member = tmqAddObjectToArray(tags);
-        RAW_NULL_CHECK(member);
-
-        SUpdatedTagVal* pTagVal = taosArrayGet(vAlterTbReq.pMultiTag, i);
-        ADD_TO_JSON_STRING(member, "colName", pTagVal->tagName);
-
-        if (pTagVal->regexp != NULL) {
-          ADD_TO_JSON_STRING(member, "regexp", pTagVal->regexp);
-          ADD_TO_JSON_STRING(member, "replacement", pTagVal->replacement);
-        } else {
-          bool isNull = pTagVal->isNull;
-          if (!isNull) {
-            int64_t bufSize = 0;
-            if (pTagVal->tagType == TSDB_DATA_TYPE_VARBINARY) {
-              bufSize = pTagVal->nTagVal * 2 + 2 + 3;
-            } else {
-              bufSize = pTagVal->nTagVal + 3;
-            }
-            buf1 = taosMemoryCalloc(bufSize, 1);
-            RAW_NULL_CHECK(buf1);
-            code = dataConverToStr(buf1, bufSize, pTagVal->tagType, pTagVal->pTagVal, pTagVal->nTagVal, NULL);
-            if (code != TSDB_CODE_SUCCESS) {
-              uError("convert tag value to string failed");
-              goto end;
-            }
-            ADD_TO_JSON_STRING(member, "colValue", buf1);
-            taosMemoryFreeClear(buf1);
-          }
-          ADD_TO_JSON_BOOL(member, "colValueNull", isNull);
-        }
-      }
-
-      if (vAlterTbReq.whereLen > 0) {
-        buf1 = taosMemoryCalloc(vAlterTbReq.whereLen, 1);
-        RAW_NULL_CHECK(buf1);
-        buf2 = taosMemoryCalloc(vAlterTbReq.whereLen, 1);
-        RAW_NULL_CHECK(buf2);
-        memcpy(buf2, vAlterTbReq.where, vAlterTbReq.whereLen);
-        RAW_RETURN_CHECK(nodesMsgToNode(buf2, vAlterTbReq.whereLen, &pWhere));
-        int32_t tlen = 0;
-        RAW_RETURN_CHECK(nodesNodeToSQL(pWhere, buf1, vAlterTbReq.whereLen, &tlen));
-        if (tlen >= 1) buf1[tlen - 1] = 0;
-        ADD_TO_JSON_STRING(json, "where", buf1 + 1);
-        taosMemoryFreeClear(buf1);
-        taosMemoryFreeClear(buf2);
-      }
-      break;
-    }
-
-    case TSDB_ALTER_TABLE_UPDATE_COLUMN_COMPRESS: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      RAW_RETURN_CHECK(setCompressOption(json, vAlterTbReq.compress));
-      break;
-    }
-    case TSDB_ALTER_TABLE_ALTER_COLUMN_REF: {
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      ADD_TO_JSON_STRING(json, "refDbName", vAlterTbReq.refDbName);
-      ADD_TO_JSON_STRING(json, "refTbName", vAlterTbReq.refTbName);
-      ADD_TO_JSON_STRING(json, "refColName", vAlterTbReq.refColName);
-      break;
-    }
-    case TSDB_ALTER_TABLE_REMOVE_COLUMN_REF:{
-      ADD_TO_JSON_STRING(json, "colName", vAlterTbReq.colName);
-      break;
-    }
-    default:
-      break;
-  }
-
-end:
-  nodesDestroyNode(pWhere);
-  destroyAlterTbReq(&vAlterTbReq);
-  tDecoderClear(&decoder);
-  taosMemoryFree(buf);
-  taosMemoryFree(buf1);
-  taosMemoryFree(buf2);
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processDropSTable(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (pJson == NULL || metaRsp == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SDecoder     decoder = {0};
-  SVDropStbReq req = {0};
-  cJSON*       json = NULL;
-  int32_t      code = 0;
-  int32_t      lino = 0;
-  RAW_LOG_START
-
-  // decode
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-  tDecoderInit(&decoder, data, len);
-  RAW_RETURN_CHECK(tDecodeSVDropStbReq(&decoder, &req));
-
-  json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-  ADD_TO_JSON_STRING(json, "type", "drop");
-  ADD_TO_JSON_STRING(json, "tableType", "super");
-  ADD_TO_JSON_STRING(json, "tableName", req.name);
-
-end:
-  tDecoderClear(&decoder);
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-static int32_t processDeleteTable(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (pJson == NULL || metaRsp == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SDeleteRes req = {0};
-  SDecoder   coder = {0};
-  cJSON*     json = NULL;
-  int32_t    code = 0;
-  int32_t    lino = 0;
-  RAW_LOG_START
-
-  // decode and process req
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-
-  tDecoderInit(&coder, data, len);
-  RAW_RETURN_CHECK(tDecodeDeleteRes(&coder, &req));
-  //  getTbName(req.tableFName);
-  char sql[256] = {0};
-  (void)snprintf(sql, sizeof(sql), "delete from `%s` where `%s` >= %" PRId64 " and `%s` <= %" PRId64, req.tableFName,
-                 req.tsColName, req.skey, req.tsColName, req.ekey);
-
-  json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-  ADD_TO_JSON_STRING(json, "type", "delete");
-  ADD_TO_JSON_STRING(json, "sql", sql);
-
-end:
-  tDecoderClear(&coder);
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processDropTable(SMqMetaRsp* metaRsp, cJSON** pJson) {
-  if (pJson == NULL || metaRsp == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SDecoder         decoder = {0};
-  SVDropTbBatchReq req = {0};
-  cJSON*           json = NULL;
-  int32_t          code = 0;
-  int32_t          lino = 0;
-  RAW_LOG_START
-  // decode
-  void*   data = POINTER_SHIFT(metaRsp->metaRsp, sizeof(SMsgHead));
-  int32_t len = metaRsp->metaRspLen - sizeof(SMsgHead);
-  tDecoderInit(&decoder, data, len);
-  RAW_RETURN_CHECK(tDecodeSVDropTbBatchReq(&decoder, &req));
-
-  json = cJSON_CreateObject();
-  RAW_NULL_CHECK(json);
-  ADD_TO_JSON_STRING(json, "type", "drop");
-
-  cJSON* tableNameList = cJSON_AddArrayToObject(json, "tableNameList");
-  RAW_NULL_CHECK(tableNameList);
-
-  for (int32_t iReq = 0; iReq < req.nReqs; iReq++) {
-    SVDropTbReq* pDropTbReq = req.pReqs + iReq;
-    RAW_NULL_CHECK(tmqAddStringToArray(tableNameList, pDropTbReq->name));
-  }
-
-end:
-  tDecoderClear(&decoder);
-  *pJson = json;
-  RAW_LOG_END
-  return code;
-}
-
-#define PROCESS_TABLE_NOT_EXIST(code,tbName) \
-if (code == TSDB_CODE_PAR_TABLE_NOT_EXIST) { \
-  uWarn(LOG_ID_TAG " %s not exists, skip", LOG_ID_VALUE, tbName); \
-  code = TSDB_CODE_SUCCESS; \
-  continue; \
-} \
-RAW_RETURN_CHECK(code);
 
 static int32_t taosCreateStb(TAOS* taos, void* meta, uint32_t metaLen) {
   if (taos == NULL || meta == NULL) {
@@ -1019,7 +55,8 @@ static int32_t taosCreateStb(TAOS* taos, void* meta, uint32_t metaLen) {
   RAW_LOG_START
 
   RAW_RETURN_CHECK(buildRequest(*(int64_t*)taos, "", 0, NULL, false, &pRequest, 0));
-  uDebug(LOG_ID_TAG " create stable, meta:%p, metaLen:%d", LOG_ID_VALUE, meta, metaLen);
+  uDebug(LOG_ID_TAG " create stable, meta:%p, metaLen:%d, db:%s", LOG_ID_VALUE, meta, metaLen,
+         pRequest->pDb ? pRequest->pDb : "NULL");
   pRequest->syncQuery = true;
   if (!pRequest->pDb) {
     code = TSDB_CODE_PAR_DB_NOT_SPECIFIED;
@@ -1379,7 +416,8 @@ static int32_t taosCreateTable(TAOS* taos, void* meta, uint32_t metaLen) {
   SArray* pTagList = taosArrayInit(0, POINTER_BYTES);
   RAW_NULL_CHECK(pTagList);
   RAW_RETURN_CHECK(buildRequest(*(int64_t*)taos, "", 0, NULL, false, &pRequest, 0));
-  uDebug(LOG_ID_TAG " create table, meta:%p, metaLen:%d", LOG_ID_VALUE, meta, metaLen);
+  uDebug(LOG_ID_TAG " create table, meta:%p, metaLen:%d, db:%s", LOG_ID_VALUE, meta, metaLen,
+         pRequest->pDb ? pRequest->pDb : "NULL");
 
   pRequest->syncQuery = true;
   if (!pRequest->pDb) {
@@ -2069,7 +1107,7 @@ int taos_write_raw_block_with_fields(TAOS* taos, int rows, char* pData, const ch
 int taos_write_raw_block_with_fields_with_reqid(TAOS* taos, int rows, char* pData, const char* tbname,
                                                 TAOS_FIELD* fields, int numFields, int64_t reqid) {
   if (taos == NULL || pData == NULL || tbname == NULL) {
-    uError("invalid parameter in %s", __func__);
+    uError("invalid parameter in %s, taos:%p, pData:%p, tbname:%p", __func__, taos, pData, tbname);
     return TSDB_CODE_INVALID_PARA;
   }
   int32_t     code = TSDB_CODE_SUCCESS;
@@ -2081,7 +1119,7 @@ int taos_write_raw_block_with_fields_with_reqid(TAOS* taos, int rows, char* pDat
   SRequestObj* pRequest = NULL;
   RAW_RETURN_CHECK(buildRequest(*(int64_t*)taos, "", 0, NULL, false, &pRequest, reqid));
 
-  uDebug(LOG_ID_TAG " write raw block with field, rows:%d, pData:%p, tbname:%s, fields:%p, numFields:%d", LOG_ID_VALUE,
+  uDebug(LOG_ID_TAG " write raw block, rows:%d, pData:%p, tbname:%s, fields:%p, numFields:%d", LOG_ID_VALUE,
          rows, pData, tbname, fields, numFields);
 
   pRequest->syncQuery = true;
@@ -2107,6 +1145,8 @@ int taos_write_raw_block_with_fields_with_reqid(TAOS* taos, int rows, char* pDat
   SVgroupInfo vgData = {0};
   RAW_RETURN_CHECK(catalogGetTableHashVgroup(pCatalog, &conn, &pName, &vgData));
   RAW_RETURN_CHECK(catalogGetTableMeta(pCatalog, &conn, &pName, &pTableMeta));
+  uDebug(LOG_ID_TAG " write raw block got meta, tbname:%s, numOfColumns:%d", LOG_ID_VALUE,
+         tbname, pTableMeta->tableInfo.numOfColumns);
   RAW_RETURN_CHECK(smlInitHandle(&pQuery));
   pVgHash = taosHashInit(16, taosGetDefaultHashFunction(TSDB_DATA_TYPE_INT), true, HASH_NO_LOCK);
   RAW_NULL_CHECK(pVgHash);
@@ -2117,7 +1157,7 @@ int taos_write_raw_block_with_fields_with_reqid(TAOS* taos, int rows, char* pDat
 
   launchQueryImpl(pRequest, pQuery, true, NULL);
   code = pRequest->code;
-  uDebug(LOG_ID_TAG " write raw block with field return, msg:%s", LOG_ID_VALUE, tstrerror(code));
+  uDebug(LOG_ID_TAG " write raw block return, tbname:%s, msg:%s", LOG_ID_VALUE, tbname, tstrerror(code));
 
 end:
   taosMemoryFreeClear(pTableMeta);
@@ -2129,66 +1169,11 @@ end:
 }
 
 int taos_write_raw_block(TAOS* taos, int rows, char* pData, const char* tbname) {
-  return taos_write_raw_block_with_reqid(taos, rows, pData, tbname, 0);
+  return taos_write_raw_block_with_fields_with_reqid(taos, rows, pData, tbname, NULL, 0, 0);
 }
 
 int taos_write_raw_block_with_reqid(TAOS* taos, int rows, char* pData, const char* tbname, int64_t reqid) {
-  if (taos == NULL || pData == NULL || tbname == NULL) {
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t     code = TSDB_CODE_SUCCESS;
-  int32_t     lino = 0;
-  STableMeta* pTableMeta = NULL;
-  SQuery*     pQuery = NULL;
-  SHashObj*   pVgHash = NULL;
-
-  SRequestObj* pRequest = NULL;
-  RAW_RETURN_CHECK(buildRequest(*(int64_t*)taos, "", 0, NULL, false, &pRequest, reqid));
-
-  uDebug(LOG_ID_TAG " write raw block, rows:%d, pData:%p, tbname:%s", LOG_ID_VALUE, rows, pData, tbname);
-
-  pRequest->syncQuery = true;
-  if (!pRequest->pDb) {
-    uError(LOG_ID_TAG " %s no database selected", LOG_ID_VALUE, __func__);
-    code = TSDB_CODE_PAR_DB_NOT_SPECIFIED;
-    goto end;
-  }
-
-  SName pName = {TSDB_TABLE_NAME_T, pRequest->pTscObj->acctId, {0}, {0}};
-  tstrncpy(pName.dbname, pRequest->pDb, sizeof(pName.dbname));
-  tstrncpy(pName.tname, tbname, sizeof(pName.tname));
-
-  struct SCatalog* pCatalog = NULL;
-  RAW_RETURN_CHECK(catalogGetHandle(pRequest->pTscObj->pAppInfo->clusterId, &pCatalog));
-
-  SRequestConnInfo conn = {0};
-  conn.pTrans = pRequest->pTscObj->pAppInfo->pTransporter;
-  conn.requestId = pRequest->requestId;
-  conn.requestObjRefId = pRequest->self;
-  conn.mgmtEps = getEpSet_s(&pRequest->pTscObj->pAppInfo->mgmtEp);
-
-  SVgroupInfo vgData = {0};
-  RAW_RETURN_CHECK(catalogGetTableHashVgroup(pCatalog, &conn, &pName, &vgData));
-  RAW_RETURN_CHECK(catalogGetTableMeta(pCatalog, &conn, &pName, &pTableMeta));
-  RAW_RETURN_CHECK(smlInitHandle(&pQuery));
-  pVgHash = taosHashInit(16, taosGetDefaultHashFunction(TSDB_DATA_TYPE_INT), true, HASH_NO_LOCK);
-  RAW_NULL_CHECK(pVgHash);
-  RAW_RETURN_CHECK(
-      taosHashPut(pVgHash, (const char*)&vgData.vgId, sizeof(vgData.vgId), (char*)&vgData, sizeof(vgData)));
-  RAW_RETURN_CHECK(rawBlockBindData(pQuery, pTableMeta, pData, NULL, NULL, 0, false, NULL, 0, false));
-  RAW_RETURN_CHECK(smlBuildOutput(pQuery, pVgHash));
-
-  launchQueryImpl(pRequest, pQuery, true, NULL);
-  code = pRequest->code;
-  uDebug(LOG_ID_TAG " write raw block return, msg:%s", LOG_ID_VALUE, tstrerror(code));
-
-end:
-  taosMemoryFreeClear(pTableMeta);
-  qDestroyQuery(pQuery);
-  destroyRequest(pRequest);
-  taosHashCleanup(pVgHash);
-  RAW_LOG_END
-  return code;
+  return taos_write_raw_block_with_fields_with_reqid(taos, rows, pData, tbname, NULL, 0, reqid);
 }
 
 static void* getRawDataFromRes(void* pRetrieve) {
@@ -2312,15 +1297,11 @@ static bool needRefreshMeta(void* rawData, STableMeta* pTableMeta, SSchemaWrappe
     return false;
   }
   char* p = (char*)rawData;
-  // | version | total length | total rows | blankFill | total columns | flag seg| block group id | column schema | each
-  // column length |
-  p += sizeof(int32_t);
-  p += sizeof(int32_t);
-  p += sizeof(int32_t);
-  p += sizeof(int32_t);
-  p += sizeof(int32_t);
-  p += sizeof(uint64_t);
-  int8_t* fields = p;
+  // Raw block header layout:
+  // | version(int32) | totalLen(int32) | totalRows(int32) | blankFill(int32) | totalCols(int32) | flagSeg(uint64) | columnSchema... | colLen... |
+  int32_t rawBlockHeaderSize = sizeof(int32_t) * 5 + sizeof(uint64_t);
+  p += rawBlockHeaderSize;
+  int8_t* fields = (int8_t*)p;
 
   if (pSW->nCols != pTableMeta->tableInfo.numOfColumns) {
     return true;
@@ -2448,6 +1429,8 @@ static int32_t processCacheMeta(SHashObj* pVgHash, SHashObj* pNameHash, SHashObj
   RAW_LOG_START
   STableMeta* pTableMeta = NULL;
   tbInfo*     tmpInfo = (tbInfo*)taosHashGet(pNameHash, pName->tname, strlen(pName->tname));
+  uDebug("%s tname:%s, cacheHit:%d, retry:%d, hasCreateReq:%d", __func__,
+         pName->tname, tmpInfo != NULL, retry, pCreateReqDst != NULL);
   if (tmpInfo == NULL || retry > 0) {
     tbInfo info = {0};
 
@@ -2502,88 +1485,10 @@ end:
   return code;
 }
 
-static int32_t tmqWriteRawDataImpl(TAOS* taos, void* data, uint32_t dataLen) {
+static int32_t tmqWriteRawCommon(TAOS* taos, void* data, uint32_t dataLen,
+                                 _raw_decode_func_ decodeFunc, bool withMeta) {
   if (taos == NULL || data == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t   code = TSDB_CODE_SUCCESS;
-  int32_t   lino = 0;
-  SQuery*   pQuery = NULL;
-  SMqRspObj rspObj = {0};
-  SDecoder  decoder = {0};
-
-  SRequestObj*     pRequest = NULL;
-  SCatalog*        pCatalog = NULL;
-  SRequestConnInfo conn = {0};
-  RAW_RETURN_CHECK(buildRawRequest(taos, &pRequest, &pCatalog, &conn));
-  uDebug(LOG_ID_TAG " write raw data, data:%p, dataLen:%d", LOG_ID_VALUE, data, dataLen);
-  RAW_RETURN_CHECK(decodeRawData(&decoder, data, dataLen, tDecodeMqDataRsp, &rspObj));
-
-  SHashObj* pVgHash = NULL;
-  SHashObj* pNameHash = NULL;
-  SHashObj* pMetaHash = NULL;
-  RAW_RETURN_CHECK(getRawCache(&pVgHash, &pNameHash, &pMetaHash, taos));
-  int retry = 0;
-  while (1) {
-    RAW_RETURN_CHECK(smlInitHandle(&pQuery));
-    uDebug(LOG_ID_TAG " write raw meta data block num:%d", LOG_ID_VALUE, rspObj.dataRsp.blockNum);
-    while (++rspObj.resIter < rspObj.dataRsp.blockNum) {
-      const char* tbName = (const char*)taosArrayGetP(rspObj.dataRsp.blockTbName, rspObj.resIter);
-      RAW_NULL_CHECK(tbName);
-      SSchemaWrapper* pSW = (SSchemaWrapper*)taosArrayGetP(rspObj.dataRsp.blockSchema, rspObj.resIter);
-      RAW_NULL_CHECK(pSW);
-      void* pRetrieve = taosArrayGetP(rspObj.dataRsp.blockData, rspObj.resIter);
-      RAW_NULL_CHECK(pRetrieve);
-      void* rawData = getRawDataFromRes(pRetrieve);
-      RAW_NULL_CHECK(rawData);
-
-      uTrace(LOG_ID_TAG " write raw data block tbname:%s", LOG_ID_VALUE, tbName);
-      SName pName = {TSDB_TABLE_NAME_T, pRequest->pTscObj->acctId, {0}, {0}};
-      tstrncpy(pName.dbname, pRequest->pDb, TSDB_DB_NAME_LEN);
-      tstrncpy(pName.tname, tbName, TSDB_TABLE_NAME_LEN);
-
-      STableMeta* pTableMeta = NULL;
-      code = processCacheMeta(pVgHash, pNameHash, pMetaHash, NULL, pCatalog, &conn, &pName, &pTableMeta, pSW,
-                                        rawData, retry);
-      PROCESS_TABLE_NOT_EXIST(code, tbName)
-      char err[ERR_MSG_LEN] = {0};
-      code = rawBlockBindData(pQuery, pTableMeta, rawData, NULL, pSW, pSW->nCols, true, err, ERR_MSG_LEN, true);
-      if (code != TSDB_CODE_SUCCESS) {
-        SET_ERROR_MSG("table:%s, err:%s", pName.tname, err);
-        goto end;
-      }
-    }
-    if (taosHashGetSize(pVgHash) == 0) {
-      goto end;
-    }
-    RAW_RETURN_CHECK(smlBuildOutput(pQuery, pVgHash));
-    launchQueryImpl(pRequest, pQuery, true, NULL);
-    code = pRequest->code;
-
-    if (NEED_CLIENT_HANDLE_ERROR(code) && retry++ < 3) {
-      uInfo("write raw retry:%d/3 end code:%d, msg:%s", retry, code, tstrerror(code));
-      qDestroyQuery(pQuery);
-      pQuery = NULL;
-      rspObj.resIter = -1;
-      continue;
-    }
-    break;
-  }
-  uDebug(LOG_ID_TAG " write raw data return, msg:%s", LOG_ID_VALUE, tstrerror(code));
-
-end:
-  tDeleteMqDataRsp(&rspObj.dataRsp);
-  tDecoderClear(&decoder);
-  qDestroyQuery(pQuery);
-  destroyRequest(pRequest);
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen) {
-  if (taos == NULL || data == NULL) {
-    uError("invalid parameter in %s", __func__);
+    uError("invalid parameter in %s, taos:%p, data:%p, withMeta:%d", __func__, taos, data, withMeta);
     return TSDB_CODE_INVALID_PARA;
   }
   int32_t   code = TSDB_CODE_SUCCESS;
@@ -2596,14 +1501,15 @@ static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen)
   SRequestObj*     pRequest = NULL;
   SCatalog*        pCatalog = NULL;
   SRequestConnInfo conn = {0};
-
   RAW_RETURN_CHECK(buildRawRequest(taos, &pRequest, &pCatalog, &conn));
-  uDebug(LOG_ID_TAG " write raw metadata, data:%p, dataLen:%d", LOG_ID_VALUE, data, dataLen);
-  RAW_RETURN_CHECK(decodeRawData(&decoder, data, dataLen, tDecodeSTaosxRsp, &rspObj));
+  uDebug(LOG_ID_TAG " write raw %s, data:%p, dataLen:%d", LOG_ID_VALUE, withMeta ? "metadata" : "data", data, dataLen);
+  RAW_RETURN_CHECK(decodeRawData(&decoder, data, dataLen, decodeFunc, &rspObj));
 
-  pCreateTbHash = taosHashInit(16, taosGetDefaultHashFunction(TSDB_DATA_TYPE_BINARY), true, HASH_NO_LOCK);
-  RAW_NULL_CHECK(pCreateTbHash);
-  RAW_RETURN_CHECK(buildCreateTbMap(&rspObj.dataRsp, pCreateTbHash));
+  if (withMeta) {
+    pCreateTbHash = taosHashInit(16, taosGetDefaultHashFunction(TSDB_DATA_TYPE_BINARY), true, HASH_NO_LOCK);
+    RAW_NULL_CHECK(pCreateTbHash);
+    RAW_RETURN_CHECK(buildCreateTbMap(&rspObj.dataRsp, pCreateTbHash));
+  }
 
   SHashObj* pVgHash = NULL;
   SHashObj* pNameHash = NULL;
@@ -2612,7 +1518,8 @@ static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen)
   int retry = 0;
   while (1) {
     RAW_RETURN_CHECK(smlInitHandle(&pQuery));
-    uDebug(LOG_ID_TAG " write raw meta data block num:%d", LOG_ID_VALUE, rspObj.dataRsp.blockNum);
+    uDebug(LOG_ID_TAG " write raw %s block num:%d, retry:%d", LOG_ID_VALUE,
+           withMeta ? "metadata" : "data", rspObj.dataRsp.blockNum, retry);
     while (++rspObj.resIter < rspObj.dataRsp.blockNum) {
       const char* tbName = (const char*)taosArrayGetP(rspObj.dataRsp.blockTbName, rspObj.resIter);
       RAW_NULL_CHECK(tbName);
@@ -2623,21 +1530,24 @@ static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen)
       void* rawData = getRawDataFromRes(pRetrieve);
       RAW_NULL_CHECK(rawData);
 
-      uTrace(LOG_ID_TAG " write raw data block tbname:%s", LOG_ID_VALUE, tbName);
+      uTrace(LOG_ID_TAG " write raw %s block[%d] tbname:%s, schemaCols:%d", LOG_ID_VALUE,
+             withMeta ? "metadata" : "data", rspObj.resIter, tbName, pSW->nCols);
       SName pName = {TSDB_TABLE_NAME_T, pRequest->pTscObj->acctId, {0}, {0}};
       tstrncpy(pName.dbname, pRequest->pDb, TSDB_DB_NAME_LEN);
       tstrncpy(pName.tname, tbName, TSDB_TABLE_NAME_LEN);
 
-      // find schema data info
-      SVCreateTbReq* pCreateReqDst = (SVCreateTbReq*)taosHashGet(pCreateTbHash, pName.tname, strlen(pName.tname));
-      STableMeta*    pTableMeta = NULL;
+      SVCreateTbReq* pCreateReqDst = NULL;
+      if (pCreateTbHash) {
+        pCreateReqDst = (SVCreateTbReq*)taosHashGet(pCreateTbHash, pName.tname, strlen(pName.tname));
+      }
+      STableMeta* pTableMeta = NULL;
       code = processCacheMeta(pVgHash, pNameHash, pMetaHash, pCreateReqDst, pCatalog, &conn, &pName,
-                                        &pTableMeta, pSW, rawData, retry);
+                              &pTableMeta, pSW, rawData, retry);
       PROCESS_TABLE_NOT_EXIST(code, tbName)
       char err[ERR_MSG_LEN] = {0};
-      code =
-          rawBlockBindData(pQuery, pTableMeta, rawData, pCreateReqDst, pSW, pSW->nCols, true, err, ERR_MSG_LEN, true);
+      code = rawBlockBindData(pQuery, pTableMeta, rawData, pCreateReqDst, pSW, pSW->nCols, true, err, ERR_MSG_LEN, true);
       if (code != TSDB_CODE_SUCCESS) {
+        uError(LOG_ID_TAG " rawBlockBindData failed, table:%s, err:%s, code:%s", LOG_ID_VALUE, pName.tname, err, tstrerror(code));
         SET_ERROR_MSG("table:%s, err:%s", pName.tname, err);
         goto end;
       }
@@ -2650,7 +1560,7 @@ static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen)
     code = pRequest->code;
 
     if (NEED_CLIENT_HANDLE_ERROR(code) && retry++ < 3) {
-      uInfo("write raw retry:%d/3 end code:%d, msg:%s", retry, code, tstrerror(code));
+      uInfo(LOG_ID_TAG " write raw retry:%d/3, code:%d, msg:%s", LOG_ID_VALUE, retry, code, tstrerror(code));
       qDestroyQuery(pQuery);
       pQuery = NULL;
       rspObj.resIter = -1;
@@ -2658,21 +1568,33 @@ static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen)
     }
     break;
   }
-  uDebug(LOG_ID_TAG " write raw metadata return, msg:%s", LOG_ID_VALUE, tstrerror(code));
+  uDebug(LOG_ID_TAG " write raw %s return, msg:%s", LOG_ID_VALUE, withMeta ? "metadata" : "data", tstrerror(code));
 
 end:
-  tDeleteSTaosxRsp(&rspObj.dataRsp);
-  void* pIter = taosHashIterate(pCreateTbHash, NULL);
-  while (pIter) {
-    tDestroySVCreateTbReq(pIter, TSDB_MSG_FLG_DECODE);
-    pIter = taosHashIterate(pCreateTbHash, pIter);
+  if (withMeta) {
+    tDeleteSTaosxRsp(&rspObj.dataRsp);
+    void* pIter = taosHashIterate(pCreateTbHash, NULL);
+    while (pIter) {
+      tDestroySVCreateTbReq(pIter, TSDB_MSG_FLG_DECODE);
+      pIter = taosHashIterate(pCreateTbHash, pIter);
+    }
+    taosHashCleanup(pCreateTbHash);
+  } else {
+    tDeleteMqDataRsp(&rspObj.dataRsp);
   }
-  taosHashCleanup(pCreateTbHash);
   tDecoderClear(&decoder);
   qDestroyQuery(pQuery);
   destroyRequest(pRequest);
   RAW_LOG_END
   return code;
+}
+
+static int32_t tmqWriteRawDataImpl(TAOS* taos, void* data, uint32_t dataLen) {
+  return tmqWriteRawCommon(taos, data, dataLen, tDecodeMqDataRsp, false);
+}
+
+static int32_t tmqWriteRawMetaDataImpl(TAOS* taos, void* data, uint32_t dataLen) {
+  return tmqWriteRawCommon(taos, data, dataLen, tDecodeSTaosxRsp, true);
 }
 
 static int32_t tmqWriteRawRawDataImpl(TAOS* taos, void* data, uint32_t dataLen) {
@@ -2762,111 +1684,6 @@ end:
   return code;
 }
 
-static int32_t processSimpleMeta(SMqMetaRsp* pMetaRsp, cJSON** meta) {
-  if (pMetaRsp == NULL || meta == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  int32_t code = 0;
-  int32_t lino = 0;
-  RAW_LOG_START
-  if (pMetaRsp->resMsgType == TDMT_VND_CREATE_STB) {
-    RAW_RETURN_CHECK(processCreateStb(pMetaRsp, meta));
-  } else if (pMetaRsp->resMsgType == TDMT_VND_ALTER_STB) {
-    RAW_RETURN_CHECK(processAlterStb(pMetaRsp, meta));
-  } else if (pMetaRsp->resMsgType == TDMT_VND_DROP_STB) {
-    RAW_RETURN_CHECK(processDropSTable(pMetaRsp, meta));
-  } else if (pMetaRsp->resMsgType == TDMT_VND_CREATE_TABLE) {
-    RAW_RETURN_CHECK(processCreateTable(pMetaRsp, meta));
-  } else if (pMetaRsp->resMsgType == TDMT_VND_ALTER_TABLE) {
-    RAW_RETURN_CHECK(processAlterTable(pMetaRsp, meta));
-  } else if (pMetaRsp->resMsgType == TDMT_VND_DROP_TABLE) {
-    RAW_RETURN_CHECK(processDropTable(pMetaRsp, meta));
-  } else if (pMetaRsp->resMsgType == TDMT_VND_DELETE) {
-    RAW_RETURN_CHECK(processDeleteTable(pMetaRsp, meta));
-  }
-
-end:
-  RAW_LOG_END
-  return code;
-}
-
-static int32_t processBatchMetaToJson(SMqBatchMetaRsp* pMsgRsp, char** string) {
-  if (pMsgRsp == NULL || string == NULL) {
-    uError("invalid parameter in %s", __func__);
-    return TSDB_CODE_INVALID_PARA;
-  }
-  SDecoder        coder = {0};
-  SMqBatchMetaRsp rsp = {0};
-  int32_t         code = 0;
-  int32_t         lino = 0;
-  cJSON*          pJson = NULL;
-  tDecoderInit(&coder, pMsgRsp->pMetaBuff, pMsgRsp->metaBuffLen);
-  RAW_RETURN_CHECK(tDecodeMqBatchMetaRsp(&coder, &rsp));
-
-  pJson = cJSON_CreateObject();
-  RAW_NULL_CHECK(pJson);
-  RAW_FALSE_CHECK(cJSON_AddStringToObject(pJson, "tmq_meta_version", TMQ_META_VERSION));
-  cJSON* pMetaArr = cJSON_AddArrayToObject(pJson, "metas");
-  RAW_NULL_CHECK(pMetaArr);
-
-  int32_t num = taosArrayGetSize(rsp.batchMetaReq);
-  for (int32_t i = 0; i < num; i++) {
-    int32_t* len = taosArrayGet(rsp.batchMetaLen, i);
-    RAW_NULL_CHECK(len);
-    void* tmpBuf = taosArrayGetP(rsp.batchMetaReq, i);
-    RAW_NULL_CHECK(tmpBuf);
-    SDecoder   metaCoder = {0};
-    SMqMetaRsp metaRsp = {0};
-    tDecoderInit(&metaCoder, POINTER_SHIFT(tmpBuf, sizeof(SMqRspHead)), *len - sizeof(SMqRspHead));
-    RAW_RETURN_CHECK(tDecodeMqMetaRsp(&metaCoder, &metaRsp));
-    cJSON* pItem = NULL;
-    RAW_RETURN_CHECK(processSimpleMeta(&metaRsp, &pItem));
-    tDeleteMqMetaRsp(&metaRsp);
-    if (pItem != NULL) RAW_FALSE_CHECK(cJSON_AddItemToArray(pMetaArr, pItem));
-  }
-
-  char* fullStr = cJSON_PrintUnformatted(pJson);
-  *string = fullStr;
-
-end:
-  cJSON_Delete(pJson);
-  tDeleteMqBatchMetaRsp(&rsp);
-  RAW_LOG_END
-  return code;
-}
-
-char* tmq_get_json_meta(TAOS_RES* res) {
-  int32_t code = TSDB_CODE_SUCCESS;
-  int32_t lino = 0;
-  char*   string = NULL;
-  RAW_LOG_START
-  RAW_NULL_CHECK(res);
-  RAW_FALSE_CHECK(TD_RES_TMQ_META(res) || TD_RES_TMQ_METADATA(res) || TD_RES_TMQ_BATCH_META(res));
-
-  SMqRspObj* rspObj = (SMqRspObj*)res;
-  if (TD_RES_TMQ_METADATA(res)) {
-    RAW_RETURN_CHECK(processAutoCreateTable(&rspObj->dataRsp, &string));
-  } else if (TD_RES_TMQ_BATCH_META(res)) {
-    RAW_RETURN_CHECK(processBatchMetaToJson(&rspObj->batchMetaRsp, &string));
-  } else if (TD_RES_TMQ_META(res)) {
-    cJSON* pJson = NULL;
-    RAW_RETURN_CHECK(processSimpleMeta(&rspObj->metaRsp, &pJson));
-    string = cJSON_PrintUnformatted(pJson);
-    cJSON_Delete(pJson);
-  } else {
-    uError("tmq_get_json_meta res:%d, invalid type", *(int8_t*)res);
-  }
-
-  uDebug("tmq_get_json_meta string:%s", string);
-
-end:
-  RAW_LOG_END
-  return string;
-}
-
-void tmq_free_json_meta(char* jsonMeta) { taosMemoryFreeClear(jsonMeta); }
-
 static int32_t getOffSetLen(const SMqDataRsp* pRsp) {
   if (pRsp == NULL) {
     uError("invalid parameter in %s", __func__);
@@ -2935,6 +1752,7 @@ int32_t tmq_get_raw(TAOS_RES* res, tmq_raw_data* raw) {
   RAW_LOG_START
   *raw = (tmq_raw_data){0};
   SMqRspObj* rspObj = ((SMqRspObj*)res);
+  uDebug("tmq_get_raw resType:%d", rspObj->resType);
   if (TD_RES_TMQ_META(res)) {
     raw->raw = rspObj->metaRsp.metaRsp;
     raw->raw_len = rspObj->metaRsp.metaRspLen >= 0 ? rspObj->metaRsp.metaRspLen : 0;
@@ -3008,30 +1826,33 @@ static int32_t writeRawImpl(TAOS* taos, void* buf, uint32_t len, uint16_t type) 
     return TSDB_CODE_INTERNAL_ERROR;
   }
 
-  if (type == TDMT_VND_CREATE_STB) {
-    return taosCreateStb(taos, buf, len);
-  } else if (type == TDMT_VND_ALTER_STB) {
-    return taosCreateStb(taos, buf, len);
-  } else if (type == TDMT_VND_DROP_STB) {
-    return taosDropStb(taos, buf, len);
-  } else if (type == TDMT_VND_CREATE_TABLE) {
-    return taosCreateTable(taos, buf, len);
-  } else if (type == TDMT_VND_ALTER_TABLE) {
-    return taosAlterTable(taos, buf, len);
-  } else if (type == TDMT_VND_DROP_TABLE) {
-    return taosDropTable(taos, buf, len);
-  } else if (type == TDMT_VND_DELETE) {
-    return taosDeleteData(taos, buf, len);
-  } else if (type == RES_TYPE__TMQ_METADATA) {
-    return tmqWriteRawMetaDataImpl(taos, buf, len);
-  } else if (type == RES_TYPE__TMQ_RAWDATA) {
-    return tmqWriteRawRawDataImpl(taos, buf, len);
-  } else if (type == RES_TYPE__TMQ) {
-    return tmqWriteRawDataImpl(taos, buf, len);
-  } else if (type == RES_TYPE__TMQ_BATCH_META) {
-    return tmqWriteBatchMetaDataImpl(taos, buf, len);
+  uDebug("writeRawImpl type:%d, buf:%p, len:%d", type, buf, len);
+  switch (type) {
+    case TDMT_VND_CREATE_STB:
+    case TDMT_VND_ALTER_STB:
+      return taosCreateStb(taos, buf, len);
+    case TDMT_VND_DROP_STB:
+      return taosDropStb(taos, buf, len);
+    case TDMT_VND_CREATE_TABLE:
+      return taosCreateTable(taos, buf, len);
+    case TDMT_VND_ALTER_TABLE:
+      return taosAlterTable(taos, buf, len);
+    case TDMT_VND_DROP_TABLE:
+      return taosDropTable(taos, buf, len);
+    case TDMT_VND_DELETE:
+      return taosDeleteData(taos, buf, len);
+    case RES_TYPE__TMQ_METADATA:
+      return tmqWriteRawMetaDataImpl(taos, buf, len);
+    case RES_TYPE__TMQ_RAWDATA:
+      return tmqWriteRawRawDataImpl(taos, buf, len);
+    case RES_TYPE__TMQ:
+      return tmqWriteRawDataImpl(taos, buf, len);
+    case RES_TYPE__TMQ_BATCH_META:
+      return tmqWriteBatchMetaDataImpl(taos, buf, len);
+    default:
+      uError("writeRawImpl unknown type:%d", type);
+      return TSDB_CODE_INVALID_PARA;
   }
-  return TSDB_CODE_INVALID_PARA;
 }
 
 int32_t tmq_write_raw(TAOS* taos, tmq_raw_data raw) {
@@ -3040,8 +1861,13 @@ int32_t tmq_write_raw(TAOS* taos, tmq_raw_data raw) {
     return TSDB_CODE_INVALID_PARA;
   }
   taosClearErrMsg();  // clear global error message
+  uDebug("tmq_write_raw connId:0x%" PRIx64 ", raw_type:%d, raw_len:%d", *(int64_t*)taos, raw.raw_type, raw.raw_len);
 
-  return writeRawImpl(taos, raw.raw, raw.raw_len, raw.raw_type);
+  int32_t code = writeRawImpl(taos, raw.raw, raw.raw_len, raw.raw_type);
+  if (code != TSDB_CODE_SUCCESS) {
+    uError("tmq_write_raw connId:0x%" PRIx64 " failed, raw_type:%d, code:%s", *(int64_t*)taos, raw.raw_type, tstrerror(code));
+  }
+  return code;
 }
 
 static int32_t tmqWriteBatchMetaDataImpl(TAOS* taos, void* meta, uint32_t metaLen) {
@@ -3059,6 +1885,7 @@ static int32_t tmqWriteBatchMetaDataImpl(TAOS* taos, void* meta, uint32_t metaLe
   tDecoderInit(&coder, meta, metaLen);
   RAW_RETURN_CHECK(tDecodeMqBatchMetaRsp(&coder, &rsp));
   int32_t num = taosArrayGetSize(rsp.batchMetaReq);
+  uDebug("%s batch meta count:%d, metaLen:%d", __func__, num, metaLen);
   for (int32_t i = 0; i < num; i++) {
     int32_t* len = taosArrayGet(rsp.batchMetaLen, i);
     RAW_NULL_CHECK(len);
@@ -3068,9 +1895,13 @@ static int32_t tmqWriteBatchMetaDataImpl(TAOS* taos, void* meta, uint32_t metaLe
     SMqMetaRsp metaRsp = {0};
     tDecoderInit(&metaCoder, POINTER_SHIFT(tmpBuf, sizeof(SMqRspHead)), *len - sizeof(SMqRspHead));
     RAW_RETURN_CHECK(tDecodeMqMetaRsp(&metaCoder, &metaRsp));
+    uDebug("%s processing batch item %d/%d, resMsgType:%d, metaRspLen:%d", __func__, i, num,
+           metaRsp.resMsgType, metaRsp.metaRspLen);
     code = writeRawImpl(taos, metaRsp.metaRsp, metaRsp.metaRspLen, metaRsp.resMsgType);
     tDeleteMqMetaRsp(&metaRsp);
     if (code != TSDB_CODE_SUCCESS) {
+      uError("%s batch item %d/%d failed, resMsgType:%d, code:%s", __func__, i, num,
+             metaRsp.resMsgType, tstrerror(code));
       goto end;
     }
   }
