+++ source/dnode/vnode/src/meta/metaQuery.c
@@ -503,7 +503,7 @@ _err:
   return NULL;
 }
 
-static int64_t getIdInner(SMeta *pMeta, int64_t uid){
+static int64_t metaGetSuidByUid(SMeta *pMeta, int64_t uid){
   int32_t code = 0;
   int64_t id = uid;
   void   *pKey = NULL;
@@ -646,7 +646,7 @@ SSchemaWrapper *metaGetTableSchema(SMeta *pMeta, tb_uid_t uid, int32_t sver, int
 _query:
   code = tdbTbGet(pMeta->pUidIdx, &uid, sizeof(uid), &pData, &nData);
   if (code == TSDB_CODE_NOT_FOUND && ignoreExist){
-    int64_t id = getIdInner(pMeta, uid);
+    int64_t id = metaGetSuidByUid(pMeta, uid);
     pSchema = getSchemaInner(pMeta, id, sver, extSchema);
     goto _exit;
   }
